ecommerce-platform/ecommerce-bot-ia/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
aiosqlite
psycopg2-binary
pydantic[email]
python-dotenv
google-generativeai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
openai
httpx
cryptography
